Started Refactoring Primitives
