+++ src/firmware/main/app/primitives/direct_velocity.c
@@ -1,30 +1,13 @@
 #include "direct_velocity.h"
 
-#include <unused.h>
-
-#include "app/control.h"
+#include "firmware/main/app/control.h"
 
 static float direct_target_velocity_x;
 static float direct_target_velocity_y;
 static float direct_target_velocity_angular;
 
-/**
- * \brief Initializes the direct_velocity primitive.
- *
- * This function runs once at system startup.
- */
 static void direct_velocity_init(void) {}
 
-/**
- * \brief Starts a movement of this type.
- *
- * This function runs each time the host computer requests to start a direct_velocity
- * movement.
- *
- * \param[in] params the movement parameters, which are only valid until this
- * function returns and must be copied into this module if needed
- * \param[in] world The world to perform the primitive in
- */
 static void direct_velocity_start(const primitive_params_t* params,
                                   FirmwareWorld_t* world)
 {
@@ -37,25 +20,9 @@ static void direct_velocity_start(const primitive_params_t* params,
     app_dribbler_setSpeed(dribbler, (params->extra) * 300);
 }
 
-/**
- * \brief Ends a movement of this type.
- *
- * This function runs when the host computer requests a new movement while a
- * direct_velocity movement is already in progress.
- * \param[in] world The world to perform the primitive in
- */
 static void direct_velocity_end(FirmwareWorld_t* world) {}
 
-/**
- * \brief Ticks a movement of this type.
- *
- * This function runs at the system tick rate while this primitive is active.
- *
- * \param[out] log the log record to fill with information about the tick, or
- * \c NULL if no record is to be filled
- * \param[in] world an object representing the world
- */
-static void direct_velocity_tick(log_record_t* UNUSED(log), FirmwareWorld_t* world)
+static void direct_velocity_tick(FirmwareWorld_t* world)
 {
     FirmwareRobot_t* robot = app_firmware_world_getRobot(world);
     app_control_trackVelocity(robot, direct_target_velocity_x, direct_target_velocity_y,
