--- src/firmware/main/app/primitives/primitive.c
@@ -1,189 +1,0 @@
-#ifndef FWSIM
-#include <FreeRTOS.h>
-#include <semphr.h>
-
-#include "imu_test.h"
-#endif  // FWSIM
-
-#include <assert.h>
-#include <stdint.h>
-#include <stdio.h>
-
-#include "catch.h"
-#include "direct_velocity.h"
-#include "direct_wheels.h"
-#include "dribble.h"
-#include "io/dr.h"
-#include "io/receive.h"
-#include "move.h"
-#include "pivot.h"
-#include "primitive.h"
-#include "shoot.h"
-#include "spin.h"
-#include "stop.h"
-
-
-// need the ifndef here so that we can ignore this code when compiling
-// the firmware tests
-#ifndef FWTEST
-/**
- * \brief The available movement primitives.
- *
- * This array is indexed by movement primitive number. If movement primitives
- * are added or removed, they must be added or removed in this array and it
- * must be kept in the same order as the enumeration in @c
- * software/mrf/constants.h.
- *
- * Make sure stop is always the first primitive.
- */
-static const primitive_t *const PRIMITIVES[] = {
-    &STOP_PRIMITIVE,     // index 0, do not change the order of stuff in this one
-    &MOVE_PRIMITIVE,     // 1
-    &DRIBBLE_PRIMITIVE,  // 2 this should be unessisary for FWSIM, but keep it here to
-                         // make the order still correct
-    &SHOOT_PRIMITIVE,    // 3
-    &CATCH_PRIMITIVE,    // 4
-    &PIVOT_PRIMITIVE,    // 5
-    &SPIN_PRIMITIVE,     // 6
-    &DIRECT_WHEELS_PRIMITIVE,    // 7
-    &DIRECT_VELOCITY_PRIMITIVE,  // 8
-    &IMU_TEST_PRIMITIVE,         // 9
-};
-
-/**
- * \brief The number of primitives.
- */
-#define PRIMITIVE_COUNT (sizeof(PRIMITIVES) / sizeof(*PRIMITIVES))
-
-/**
- * \brief The mutex that prevents multiple entries into the same primitive at
- * the same time.
- */
-#ifndef FWSIM
-static SemaphoreHandle_t primitive_mutex;
-#endif  // FWSIM
-
-/**
- * \brief The primitive that is currently operating.
- */
-static const primitive_t *primitive_current;
-
-/**
- * \brief The index number of the current primitive.
- */
-static unsigned int primitive_current_index;
-
-/**
- * \brief Initializes the movement primitive manager and all the primitives.
- */
-void primitive_init(void)
-{
-#ifndef FWSIM
-    static StaticSemaphore_t primitive_mutex_storage;
-    primitive_mutex = xSemaphoreCreateMutexStatic(&primitive_mutex_storage);
-#endif  // FWSIM
-    for (size_t i = 0; i != PRIMITIVE_COUNT; ++i)
-    {
-        PRIMITIVES[i]->init();
-    }
-}
-
-/**
- * \brief Starts a new movement.
- *
- * \param[in] primitive the index of the primitive to run
- * \param[in] params the parameters to the primitive
- * \param[in] world The world to perform the primitive in
- */
-void primitive_start(unsigned int primitive, const primitive_params_t *params,
-                     FirmwareWorld_t *world)
-{
-    assert(primitive < PRIMITIVE_COUNT);
-    xSemaphoreTake(primitive_mutex, portMAX_DELAY);
-    if (primitive_current)
-    {
-        primitive_current->end(world);
-    }
-
-    FirmwareRobot_t *robot = app_firmware_world_getRobot(world);
-    Chicker_t *chicker     = app_firmware_robot_getChicker(robot);
-    Dribbler_t *dribbler   = app_firmware_robot_getDribbler(robot);
-
-    app_chicker_disableAutochip(chicker);
-    app_chicker_disableAutokick(chicker);
-    app_dribbler_setSpeed(dribbler, 0);
-    primitive_current       = PRIMITIVES[primitive];
-    primitive_current_index = primitive;
-    primitive_current->start(params, world);
-    xSemaphoreGive(primitive_mutex);
-}
-
-/**
- * \brief Ticks the current primitive.
- *
- * \param[out] log the log record to fill with information about the tick, or
- * \c NULL if no record is to be filled
- * \param[in] world An object representing the world
- */
-void primitive_tick(log_record_t *log, FirmwareWorld_t *world)
-{
-    xSemaphoreTake(primitive_mutex, portMAX_DELAY);
-    if (log)
-    {
-        log->tick.drive_serial = receive_last_serial();
-        log->tick.primitive    = (uint8_t)primitive_current_index;
-    }
-    if (primitive_current)
-    {
-        primitive_current->tick(log, world);
-    }
-    dr_tick(log);
-    xSemaphoreGive(primitive_mutex);
-}
-
-/**
- * \brief Checks whether a particular primitive is direct.
- *
- * \param[in] primitive the primitive to check
- * \retval true the primitive is direct
- * \retval false the primitive is a movement primitive
- */
-bool primitive_is_direct(unsigned int primitive)
-{
-    return PRIMITIVES[primitive]->direct;
-}
-
-unsigned int get_primitive_index()
-{
-    return primitive_current_index;
-}
-
-bool primitive_params_are_equal(primitive_params_t *params1, primitive_params_t *params2)
-{
-    bool equal = true;
-
-    if (params1->slow != params2->slow)
-    {
-        equal = false;
-    }
-
-    if (equal && params1->extra != params2->extra)
-    {
-        equal = false;
-    }
-
-    if (equal)
-    {
-        for (int i = 0; i < 4; i++)
-        {
-            if (params1->params[i] != params2->params[i])
-            {
-                equal = false;
-            }
-        }
-    }
-
-    return equal;
-}
-
-#endif
