+++ src/firmware/main/app/primitives/stop.h
@@ -1,7 +1,7 @@
 #ifndef PRIMITIVES_STOP_H
 #define PRIMITIVES_STOP_H
 
-#include "primitive.h"
+#include "firmware/main/app/primitives/primitive.h"
 
 extern const primitive_t STOP_PRIMITIVE;
