+++ src/firmware/main/app/world/BUILD
@@ -1,13 +1,9 @@
 package(default_visibility = ["//visibility:public"])
 
-# TODO (Issue #790): remove all `strip_include_prefix` once firmware has been setup to have includes
-#       from top-level
-
 cc_library(
     name = "chicker",
     srcs = ["chicker.c"],
     hdrs = ["chicker.h"],
-    strip_include_prefix = "/firmware/main",
 )
 
 cc_test(
@@ -23,7 +19,6 @@ cc_library(
     name = "dribbler",
     srcs = ["dribbler.c"],
     hdrs = ["dribbler.h"],
-    strip_include_prefix = "/firmware/main",
 )
 
 cc_test(
@@ -39,7 +34,6 @@ cc_library(
     name = "firmware_ball",
     srcs = ["firmware_ball.c"],
     hdrs = ["firmware_ball.h"],
-    strip_include_prefix = "/firmware/main",
 )
 
 cc_test(
@@ -55,7 +49,6 @@ cc_library(
     name = "firmware_robot",
     srcs = ["firmware_robot.c"],
     hdrs = ["firmware_robot.h"],
-    strip_include_prefix = "/firmware/main",
     deps = [
         ":chicker",
         ":dribbler",
@@ -76,7 +69,6 @@ cc_library(
     name = "firmware_world",
     srcs = ["firmware_world.c"],
     hdrs = ["firmware_world.h"],
-    strip_include_prefix = "/firmware/main",
     deps = [
         ":firmware_ball",
         ":firmware_robot",
@@ -96,7 +88,6 @@ cc_library(
     name = "wheel",
     srcs = ["wheel.c"],
     hdrs = ["wheel.h"],
-    strip_include_prefix = "/firmware/main",
 )
 
 cc_test(
