+++ src/firmware/main/app/world/dribbler.c
@@ -1,4 +1,4 @@
-#include "app/world/dribbler.h"
+#include "firmware/main/app/world/dribbler.h"
 
 #include <stdlib.h>
