+++ src/firmware/main/app/world/firmware_ball.c
@@ -1,4 +1,4 @@
-#include "app/world/firmware_ball.h"
+#include "firmware/main/app/world/firmware_ball.h"
 
 #include <stdlib.h>
