+++ src/firmware/main/app/world/firmware_robot.h
@@ -1,8 +1,8 @@
 #pragma once
 
-#include "app/world/chicker.h"
-#include "app/world/dribbler.h"
-#include "app/world/wheel.h"
+#include "firmware/main/app/world/chicker.h"
+#include "firmware/main/app/world/dribbler.h"
+#include "firmware/main/app/world/wheel.h"
 
 /**
  * This struct represents a robot from the perspective of firmware
@@ -93,28 +93,28 @@ void app_firmware_robot_destroy(FirmwareRobot_t* robot);
  * @param robot The robot to get the chicker from
  * @return The chicker from the given robot
  */
-Chicker_t* app_firmware_robot_getChicker(FirmwareRobot_t* robot);
+Chicker_t* app_firmware_robot_getChicker(const FirmwareRobot_t* robot);
 
 /**
  * Get the dribbler from the given robot
  * @param robot The robot to get the dribbler from
  * @return The dribbler from the given robot
  */
-Dribbler_t* app_firmware_robot_getDribbler(FirmwareRobot_t* robot);
+Dribbler_t* app_firmware_robot_getDribbler(const FirmwareRobot_t* robot);
 
 /**
  * Get the x-position of the given robot
  * @param robot The robot to get the y-position for
  * @return The x-position of the given robot, in meters, in global coordinates
  */
-float app_firmware_robot_getPositionX(FirmwareRobot_t* robot);
+float app_firmware_robot_getPositionX(const FirmwareRobot_t* robot);
 
 /**
  * Get the y-position of the given robot
  * @param robot The robot to get the y-position for
  * @return The y-position of the given robot, in meters, in global coordinates
  */
-float app_firmware_robot_getPositionY(FirmwareRobot_t* robot);
+float app_firmware_robot_getPositionY(const FirmwareRobot_t* robot);
 
 /**
  * Get the current orientation of the given robot
@@ -122,56 +122,56 @@ float app_firmware_robot_getPositionY(FirmwareRobot_t* robot);
  * @param robot The robot to get the orientation of
  * @return The orientation of the robot, in radians
  */
-float app_firmware_robot_getOrientation(FirmwareRobot_t* robot);
+float app_firmware_robot_getOrientation(const FirmwareRobot_t* robot);
 
 /**
  * Get the x-velocity of the given robot
  * @param robot The robot to get the y-velocity for
  * @return The x-velocity of the given robot, in m/s, in global coordinates
  */
-float app_firmware_robot_getVelocityX(FirmwareRobot_t* robot);
+float app_firmware_robot_getVelocityX(const FirmwareRobot_t* robot);
 
 /**
  * Get the y-velocity of the given robot
  * @param robot The robot to get the y-velocity for
  * @return The y-velocity of the given robot, in m/s, in global coordinates
  */
-float app_firmware_robot_getVelocityY(FirmwareRobot_t* robot);
+float app_firmware_robot_getVelocityY(const FirmwareRobot_t* robot);
 
 /**
  * Get the current angular velocity for the given robot
  * @param robot The robot to get the angular velocity for
  * @return The angular velocity of the given robot, in rad/s
  */
-float app_firmware_robot_getVelocityAngular(FirmwareRobot_t* robot);
+float app_firmware_robot_getVelocityAngular(const FirmwareRobot_t* robot);
 
 /**
  * Get the battery voltage for the given robot
  * @param robot The robot to get the battery voltage for
  * @return The battery voltage for the given robot, in volts
  */
-float app_firmware_robot_getBatteryVoltage(FirmwareRobot_t* robot);
+float app_firmware_robot_getBatteryVoltage(const FirmwareRobot_t* robot);
 
 /**
  * Get the front right wheel from the given robot
  * @param robot The robot to get the front right wheel from
  * @return The front right wheel from the given robot
  */
-Wheel_t* app_firmware_robot_getFrontRightWheel(FirmwareRobot_t* robot);
+Wheel_t* app_firmware_robot_getFrontRightWheel(const FirmwareRobot_t* robot);
 
 /**
  * Get the front left wheel from the given robot
  * @param robot The robot to get the front left wheel from
  * @return The front left wheel from the given robot
  */
-Wheel_t* app_firmware_robot_getFrontLeftWheel(FirmwareRobot_t* robot);
+Wheel_t* app_firmware_robot_getFrontLeftWheel(const FirmwareRobot_t* robot);
 
 /**
  * Get the back right wheel from the given robot
  * @param robot The robot to get the back right wheel from
  * @return The back right wheel from the given robot
  */
-Wheel_t* app_firmware_robot_getBackRightWheel(FirmwareRobot_t* robot);
+Wheel_t* app_firmware_robot_getBackRightWheel(const FirmwareRobot_t* robot);
 
 
 /**
@@ -179,18 +179,18 @@ Wheel_t* app_firmware_robot_getBackRightWheel(FirmwareRobot_t* robot);
  * @param robot The robot to get the back left wheel from
  * @return The back left wheel from the given robot
  */
-Wheel_t* app_firmware_robot_getBackLeftWheel(FirmwareRobot_t* robot);
+Wheel_t* app_firmware_robot_getBackLeftWheel(const FirmwareRobot_t* robot);
 
 /**
  * Get the constants for this robot
  * @param robot The robot to get constants from
  * @return The constants for the given robot
  */
-const RobotConstants_t app_firmware_robot_getRobotConstants(FirmwareRobot_t* robot);
+const RobotConstants_t app_firmware_robot_getRobotConstants(const FirmwareRobot_t* robot);
 
 /**
  * Get the controller state for the given robot
  * @param robot The robot to get the controller state for
  * @return A pointer to the controller state for the given robot
  */
-ControllerState_t* app_firmware_robot_getControllerState(FirmwareRobot_t* robot);
+ControllerState_t* app_firmware_robot_getControllerState(const FirmwareRobot_t* robot);
